Consolidate caching logic
